+++ app/src/loader.js
@@ -1,0 +1,209 @@
+import * as animation  from './animation'
+import * as html_utils from './html_utils'
+import * as math       from './math'
+import * as svg        from './svg'
+
+
+
+// =========================
+// === ProgressIndicator ===
+// =========================
+
+let bg_color     = "#ffffff"
+let loader_color = "#303030"
+
+export class ProgressIndicator {
+    constructor(cfg) {
+        this.dom                = html_utils.new_top_level_div()
+        this.dom.id             = 'loader'
+        this.dom.style.position = 'absolute'
+        this.dom.style.zIndex   = -1
+
+        let center = document.createElement('div')
+        center.style.width          = '100%'
+        center.style.height         = '100%'
+        center.style.display        = 'flex'
+        center.style.justifyContent = 'center'
+        center.style.alignItems     = 'center'
+        this.dom.appendChild(center)
+
+
+        let progress_bar_svg   = this.new_svg()
+        let progress_bar       = document.createElement('div')
+        progress_bar.innerHTML = progress_bar_svg
+        center.appendChild(progress_bar)
+
+        this.progress_indicator        = document.getElementById("progress_indicator")
+        this.progress_indicator_mask   = document.getElementById("progress_indicator_mask")
+        this.progress_indicator_corner = document.getElementById("progress_indicator_corner")
+
+        this.set(0)
+        this.set_opacity(0)
+
+        if(!cfg.no_animation) {
+            this.initialized = this.animate_show()
+        } else {
+            this.initialized = new Promise((resolve) => {resolve()})
+        }
+        this.destroyed = false
+    }
+
+    new_svg() {
+        let width        = 128
+        let height       = 128
+        let alpha        = 0.9
+        let inner_radius = 48
+        let outer_radius = 60
+        let mid_radius = (inner_radius + outer_radius) / 2
+        let bar_width = outer_radius - inner_radius
+
+        return svg.new_svg(width,height,`
+            <defs>
+                <g id="progress_bar">
+                    <circle fill="${loader_color}" r="${outer_radius}"                               />
+                    <circle fill="${bg_color}"     r="${inner_radius}"                               />
+                    <path   fill="${bg_color}"     opacity="${alpha}" id="progress_indicator_mask"   />
+                    <circle fill="${loader_color}" r="${bar_width/2}" id="progress_indicator_corner" />
+                    <circle fill="${loader_color}" r="${bar_width/2}" cy="-${mid_radius}"            />
+                </g>
+            </defs>
+            <g transform="translate(${width/2},${height/2})">
+                <g transform="rotate(0,0,0)" id="progress_indicator">
+                    <use xlink:href="#progress_bar"></use>
+                </g>
+            </g>
+        `)
+    }
+
+    destroy() {
+        html_utils.remove_node(this.dom)
+        this.destroyed = true
+    }
+
+    set(value) {
+        let min_angle  = 0
+        let max_angle  = 359
+        let angle_span = max_angle - min_angle
+        let mask_angle = (1-value)*angle_span - min_angle
+        let corner_pos = math.polar_to_cartesian(54, -mask_angle)
+        this.progress_indicator_mask.setAttribute("d", svg.arc(128, -mask_angle))
+        this.progress_indicator_corner.setAttribute("cx", corner_pos.x)
+        this.progress_indicator_corner.setAttribute("cy", corner_pos.y)
+    }
+
+    set_opacity(val) {
+        this.progress_indicator.setAttribute("opacity",val)
+    }
+
+    set_rotation(val) {
+        this.progress_indicator.setAttribute("transform",`rotate(${val},0,0)`)
+    }
+
+    animate_show() {
+        let indicator = this
+        return new Promise(function(resolve, reject) {
+            let alpha = 0
+            function show_step() {
+                if (alpha > 1) { alpha = 1 }
+                indicator.set_opacity(animation.ease_in_out_quad(alpha))
+                alpha += 0.02
+                if (alpha < 1) {
+                    window.requestAnimationFrame(show_step)
+                } else {
+                     console.log("SHOW END")
+                     resolve()
+                }
+            }
+            window.requestAnimationFrame(show_step)
+        })
+    }
+
+    animate_rotation() {
+        let indicator = this
+        let rotation  = 0
+        function rotate_step(timestamp) {
+            indicator.set_rotation(rotation)
+            rotation += 6
+            if (!indicator.destroyed) {
+                window.requestAnimationFrame(rotate_step)
+            }
+        }
+        window.requestAnimationFrame(rotate_step)
+    }
+}
+
+
+
+// ==============
+// === Loader ===
+// ==============
+
+export class Loader {
+    constructor(total_bytes, cfg) {
+        this.indicator         = new ProgressIndicator(cfg)
+        this.total_bytes       = total_bytes
+        this.received_bytes    = 0
+        this.download_speed    = 0
+        this.last_receive_time = performance.now()
+        this.on_done = () => {}
+        this.initialized = this.indicator.initialized
+    }
+
+    value() {
+        if (this.total_bytes == 0) {
+            return 0.3
+        } else {
+            return this.received_bytes / this.total_bytes
+        }
+    }
+
+    done() {
+        return this.received_bytes == this.total_bytes
+    }
+
+    destroy() {
+        this.indicator.destroy()
+    }
+
+    on_receive(new_bytes) {
+        this.received_bytes += new_bytes
+        let time      = performance.now()
+        let time_diff = time - this.last_receive_time
+        this.download_speed = new_bytes / time_diff
+        this.last_receive_time = time
+
+        let percent  = this.show_percentage_value()
+        let speed    = this.show_download_speed()
+        let received = this.show_received_bytes()
+        console.log(`${percent}% (${received}) (${speed}).`)
+
+        this.indicator.set(this.value())
+
+        if (this.done()) { this.on_done() }
+    }
+
+    show_percentage_value() {
+        return Math.round(100 * this.value())
+    }
+
+    show_total_bytes() {
+        return `${math.format_mb(this.total_bytes)} MB`
+    }
+
+    show_received_bytes() {
+        return `${math.format_mb(this.received_bytes)} MB`
+    }
+
+    show_download_speed() {
+        return `${math.format_mb(1000 * this.download_speed)} MB/s`
+    }
+
+    input_stream() {
+        let loader = this
+        return new WritableStream({
+             write(t) {
+                 loader.on_receive(t.length)
+             }
+        })
+    }
+}
